requirements.txt
# Twitter Bot Requirements

# Twitter automation
twikit==2.3.3

# HTTP client
httpx==0.28.1

# Configuration
PyYAML==6.0.2

# Database
# SQLite is built-in Python

# Web Dashboard
Flask==3.1.0
Flask-CORS==5.0.0

# Utilities
python-dotenv==1.2.1
schedule==1.2.2
